# backend/src/main/java/com/wardanger/excalibur/product/domain/ProductDefinition.java
package com.wardanger.excalibur.product.domain;

import java.util.UUID;

public record ProductDefinition(
        UUID id,
        String code,
        String name,
        int validityDays,
        Integer entryLimit,
        long defaultPrice,
        boolean active) {
}