# openapi/excalibur-gym.yaml
openapi: 3.0.3
info:
  title: Excalibur Gym API
  version: 0.1.0
  description: Contract for the local Excalibur Gym administration application.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/setup/admin:
    post:
      tags:
        - Setup
      operationId: createInitialAdmin
      summary: Create the initial owner account and authenticate it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInitialAdminRequest"
      responses:
        "201":
          description: Initial administrator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedEmployee"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/employees:
    get:
      tags:
        - Employees
      operationId: listEmployees
      summary: List all employee accounts
      responses:
        "200":
          description: Employee accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - Employees
      operationId: createEmployee
      summary: Create a receptionist account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/employees/{employeeId}/status:
    patch:
      tags:
        - Employees
      operationId: updateEmployeeStatus
      summary: Activate or deactivate a receptionist account
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeStatusRequest"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/employees/{employeeId}/pin:
    put:
      tags:
        - Employees
      operationId: resetEmployeePin
      summary: Replace a receptionist PIN
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePinRequest"
      responses:
        "204":
          description: PIN replaced
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/guests:
    get:
      tags:
        - Guests
      operationId: listGuests
      summary: Search guests by name or list the most recently created guests
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
      responses:
        "200":
          description: Matching guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuestSummary"
    post:
      tags:
        - Guests
      operationId: createGuest
      summary: Register a new or imported guest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuestRequest"
      responses:
        "201":
          description: Guest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestDetails"
  /api/v1/guests/{guestId}:
    get:
      tags:
        - Guests
      operationId: getGuest
      summary: Return a guest with their pass history
      parameters:
        - $ref: "#/components/parameters/GuestId"
      responses:
        "200":
          description: Guest details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestDetails"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/guests/{guestId}/passes:
    post:
      tags:
        - Guests
      operationId: sellGuestPass
      summary: Issue a new gym pass and record its payment
      parameters:
        - $ref: "#/components/parameters/GuestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellPassRequest"
      responses:
        "201":
          description: Pass issued and payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestPass"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/guests/{guestId}/check-ins:
    post:
      tags:
        - Guests
      operationId: checkInGuest
      summary: Record a gym visit, optionally confirming a repeated same-day entry
      parameters:
        - $ref: "#/components/parameters/GuestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInGuestRequest"
      responses:
        "201":
          description: Visit recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestCheckIn"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/guests/{guestId}/check-ins/{checkInId}/reversal:
    post:
      tags:
        - Guests
      operationId: reverseGuestCheckIn
      summary: Reverse an incorrect same-day check-in while preserving its audit history
      parameters:
        - $ref: "#/components/parameters/GuestId"
        - $ref: "#/components/parameters/CheckInId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReverseCheckInRequest"
      responses:
        "200":
          description: Check-in reversed and any consumed entry restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestCheckIn"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/admin/gym-pass-products:
    get:
      tags:
        - Products
      operationId: listAllGymPassProducts
      summary: List active and inactive gym pass products for administration
      responses:
        "200":
          description: All gym pass definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductDefinition"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - Products
      operationId: createGymPassProduct
      summary: Create a configurable gym pass product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertGymPassProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDefinition"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/gym-pass-products/{productId}:
    put:
      tags:
        - Products
      operationId: updateGymPassProduct
      summary: Update a gym pass product and its current default price
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertGymPassProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDefinition"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Products
      operationId: deactivateGymPassProduct
      summary: Deactivate a product while preserving historical sales
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "204":
          description: Product deactivated
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    EmployeeId:
      name: employeeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    GuestId:
      name: guestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CheckInId:
      name: checkInId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Forbidden:
      description: Administrator permission is required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: The request conflicts with the current application state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    CreateInitialAdminRequest:
      type: object
      required:
        - displayName
        - pin
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 120
        pin:
          $ref: "#/components/schemas/Pin"
...
    AuthenticatedEmployee:
      type: object
      required:
        - id
        - displayName
        - role
        - loggedInAt
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        role:
          $ref: "#/components/schemas/EmployeeRole"
        loggedInAt:
          type: string
          format: date-time
...
    Employee:
      type: object
      required:
        - id
        - displayName
        - role
        - enabled
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        role:
          $ref: "#/components/schemas/EmployeeRole"
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateEmployeeRequest:
      type: object
      required:
        - displayName
        - pin
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 120
        pin:
          $ref: "#/components/schemas/Pin"
    UpdateEmployeeStatusRequest:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
    ChangePinRequest:
      type: object
      required:
        - pin
      properties:
        pin:
          $ref: "#/components/schemas/Pin"
    Pin:
      type: string
      pattern: "^[0-9]{4,8}$"
      description: A 4 to 8 digit employee PIN.
      example: "1234"
    EmployeeRole:
      type: string
      enum:
        - ADMIN
        - EMPLOYEE
    GuestRegistrationType:
      type: string
      enum:
        - NEW
        - IMPORTED
    PaymentMethod:
      type: string
      enum:
        - CASH
        - BANK_CARD
        - PREPAID_BALANCE
...
    CreateGuestRequest:
      type: object
      required:
        - fullName
        - registrationType
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 120
        registrationType:
          $ref: "#/components/schemas/GuestRegistrationType"
    GuestSummary:
      type: object
      required:
        - id
        - fullName
        - registrationType
        - createdAt
        - activeGymPass
        - checkedInToday
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        registrationType:
          $ref: "#/components/schemas/GuestRegistrationType"
        createdAt:
          type: string
          format: date-time
        activeGymPass:
          type: boolean
        checkedInToday:
          type: boolean
        currentPassName:
          type: string
          nullable: true
        currentPassValidUntil:
          type: string
          format: date
          nullable: true
    GuestDetails:
      allOf:
        - $ref: "#/components/schemas/GuestSummary"
        - type: object
          required:
            - passes
            - visits
          properties:
            passes:
              type: array
              items:
                $ref: "#/components/schemas/GuestPass"
            visits:
              type: array
              items:
                $ref: "#/components/schemas/GuestCheckIn"
    ProductDefinition:
      type: object
      required:
        - id
        - code
        - name
        - validityDays
        - defaultPrice
        - active
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        validityDays:
          type: integer
          format: int32
          minimum: 1
        entryLimit:
          type: integer
          format: int32
          nullable: true
          minimum: 1
        defaultPrice:
          type: integer
          format: int64
          minimum: 0
          description: Current default gross price in Hungarian forints.
        active:
          type: boolean
    UpsertGymPassProductRequest:
      type: object
      required:
        - name
        - validityDays
        - defaultPrice
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        validityDays:
          type: integer
          format: int32
          minimum: 1
          maximum: 3650
        entryLimit:
          type: integer
          format: int32
          minimum: 1
          maximum: 1000
          nullable: true
        defaultPrice:
          type: integer
          format: int64
          minimum: 0
          maximum: 10000000
    SellPassRequest:
      type: object
      required:
        - productId
        - validFrom
        - paymentMethod
      properties:
        productId:
          type: string
          format: uuid
        validFrom:
          type: string
          format: date
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
    CheckInGuestRequest:
      type: object
      required:
        - confirmRepeatedToday
      properties:
        confirmRepeatedToday:
          type: boolean
          default: false
    ReverseCheckInRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 3
          maxLength: 500
    GuestCheckIn:
      type: object
      required:
        - id
        - checkedInAt
        - passName
        - checkedInByEmployeeName
        - reversed
      properties:
        id:
          type: string
          format: uuid
        checkedInAt:
          type: string
          format: date-time
        passName:
          type: string
        checkedInByEmployeeName:
          type: string
        reversed:
          type: boolean
        reversedAt:
          type: string
          format: date-time
          nullable: true
        reversedByEmployeeName:
          type: string
          nullable: true
        reversalReason:
          type: string
          nullable: true
    GuestPass:
      type: object
      required:
        - id
        - productName
        - purchasedAt
        - validFrom
        - validUntil
        - pricePaid
        - paymentMethod
        - issuedByEmployeeName
        - status
      properties:
        id:
          type: string
          format: uuid
        productName:
          type: string
        purchasedAt:
          type: string
          format: date-time
        validFrom:
          type: string
          format: date
        validUntil:
          type: string
          format: date
        totalEntries:
          type: integer
          format: int32
          nullable: true
        remainingEntries:
          type: integer
          format: int32
          nullable: true
        pricePaid:
          type: integer
          format: int64
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        issuedByEmployeeName:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - PENDING
            - EXPIRED
            - INVALIDATED
...
    ApiError:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time